pdfsortrenamer/pdfsortrenamer/requirements.txt
pdf2image==1.17.0
pillow==11.1.0
pytesseract==0.3.13
scikit-learn==1.1.1
tqdm==4.64.0
numpy==1.23.0
